+++ packages/soulwallet-abi/src/ABI/ABI_SoulWallet.ts
@@ -7,6 +7,11 @@ export default [
                 "internalType": "contract IEntryPoint",
                 "name": "_EntryPoint",
                 "type": "address"
+            },
+            {
+                "internalType": "contract IValidator",
+                "name": "_validator",
+                "type": "address"
             }
         ],
         "stateMutability": "nonpayable",
@@ -39,6 +44,16 @@ export default [
     },
     {
         "inputs": [],
+        "name": "DATA_ALREADY_EXISTS",
+        "type": "error"
+    },
+    {
+        "inputs": [],
+        "name": "DATA_NOT_EXISTS",
+        "type": "error"
+    },
+    {
+        "inputs": [],
         "name": "HASH_ALREADY_APPROVED",
         "type": "error"
     },
@@ -50,6 +65,11 @@ export default [
     {
         "inputs": [],
         "name": "INVALID_ADDRESS",
+        "type": "error"
+    },
+    {
+        "inputs": [],
+        "name": "INVALID_DATA",
         "type": "error"
     },
     {
@@ -69,7 +89,7 @@ export default [
     },
     {
         "inputs": [],
-        "name": "INVALID_SIGNTYPE",
+        "name": "InvalidInitialization",
         "type": "error"
     },
     {
@@ -100,6 +120,11 @@ export default [
     {
         "inputs": [],
         "name": "NO_OWNER",
+        "type": "error"
+    },
+    {
+        "inputs": [],
+        "name": "NotInitializing",
         "type": "error"
     },
     {
@@ -149,11 +174,6 @@ export default [
     },
     {
         "inputs": [],
-        "name": "UNSUPPORTED_SIGNTYPE",
-        "type": "error"
-    },
-    {
-        "inputs": [],
         "name": "UPGRADE_FAILED",
         "type": "error"
     },
@@ -188,9 +208,9 @@ export default [
         "inputs": [
             {
                 "indexed": false,
-                "internalType": "uint8",
+                "internalType": "uint64",
                 "name": "version",
-                "type": "uint8"
+                "type": "uint64"
             }
         ],
         "name": "Initialized",
@@ -240,9 +260,9 @@ export default [
         "inputs": [
             {
                 "indexed": true,
-                "internalType": "address",
+                "internalType": "bytes32",
                 "name": "owner",
-                "type": "address"
+                "type": "bytes32"
             }
         ],
         "name": "OwnerAdded",
@@ -259,9 +279,9 @@ export default [
         "inputs": [
             {
                 "indexed": true,
-                "internalType": "address",
+                "internalType": "bytes32",
                 "name": "owner",
-                "type": "address"
+                "type": "bytes32"
             }
         ],
         "name": "OwnerRemoved",
@@ -358,9 +378,9 @@ export default [
     {
         "inputs": [
             {
-                "internalType": "address",
+                "internalType": "bytes32",
                 "name": "owner",
-                "type": "address"
+                "type": "bytes32"
             }
         ],
         "name": "addOwner",
@@ -371,9 +391,9 @@ export default [
     {
         "inputs": [
             {
-                "internalType": "address[]",
+                "internalType": "bytes32[]",
                 "name": "owners",
-                "type": "address[]"
+                "type": "bytes32[]"
             }
         ],
         "name": "addOwners",
@@ -523,9 +543,9 @@ export default [
     {
         "inputs": [
             {
-                "internalType": "address",
-                "name": "anOwner",
-                "type": "address"
+                "internalType": "bytes32[]",
+                "name": "owners",
+                "type": "bytes32[]"
             },
             {
                 "internalType": "address",
@@ -589,9 +609,9 @@ export default [
     {
         "inputs": [
             {
-                "internalType": "address",
-                "name": "addr",
-                "type": "address"
+                "internalType": "bytes32",
+                "name": "owner",
+                "type": "bytes32"
             }
         ],
         "name": "isOwner",
@@ -652,9 +672,9 @@ export default [
         "name": "listOwner",
         "outputs": [
             {
-                "internalType": "address[]",
+                "internalType": "bytes32[]",
                 "name": "owners",
-                "type": "address[]"
+                "type": "bytes32[]"
             }
         ],
         "stateMutability": "view",
@@ -750,9 +770,9 @@ export default [
     {
         "inputs": [
             {
-                "internalType": "address",
+                "internalType": "bytes32",
                 "name": "owner",
-                "type": "address"
+                "type": "bytes32"
             }
         ],
         "name": "removeOwner",
@@ -776,9 +796,9 @@ export default [
     {
         "inputs": [
             {
-                "internalType": "address",
+                "internalType": "bytes32",
                 "name": "newOwner",
-                "type": "address"
+                "type": "bytes32"
             }
         ],
         "name": "resetOwner",
@@ -789,9 +809,9 @@ export default [
     {
         "inputs": [
             {
-                "internalType": "address[]",
+                "internalType": "bytes32[]",
                 "name": "newOwners",
-                "type": "address[]"
+                "type": "bytes32[]"
             }
         ],
         "name": "resetOwners",
@@ -922,6 +942,19 @@ export default [
             }
         ],
         "stateMutability": "nonpayable",
+        "type": "function"
+    },
+    {
+        "inputs": [],
+        "name": "validator",
+        "outputs": [
+            {
+                "internalType": "contract IValidator",
+                "name": "",
+                "type": "address"
+            }
+        ],
+        "stateMutability": "view",
         "type": "function"
     },
     {