+++ packages/soulwallet-sdk/src/tools/gasOverhead.ts
@@ -1,3 +1,4 @@
+import { SignkeyType } from "../interface/ISoulWallet.js";
 import { UserOperation } from "../interface/UserOperation.js";
 
 
@@ -10,24 +11,46 @@ export class GasOverhead {
      * @param {UserOperation} userOp
      * @memberof GasOverhead
      */
-    public static calcGasOverhead(userOp: UserOperation) {
+    public static calcGasOverhead(userOp: UserOperation, signkeyType?: SignkeyType) {
 
-        // preVerificationGas overhead
-        {
-            // preVerificationGas overhead 1024 = 64bytes more ( 1024/16 = 64 )
-            userOp.preVerificationGas = '0x' + (BigInt(userOp.preVerificationGas) + BigInt(1024)).toString(16);
+        if (signkeyType === SignkeyType.P256) {
+            // preVerificationGas overhead
+            {
+                // preVerificationGas overhead 1024 = 64bytes more ( 1024/16 = 64 )
+                userOp.preVerificationGas = '0x' + (BigInt(userOp.preVerificationGas) + BigInt(1024)).toString(16);
+            }
+
+            // verificationGasLimit overhead
+            {
+                let verificationGasLimitOverHead = BigInt(80000);
+                if (userOp.initCode !== '0x') {
+                    verificationGasLimitOverHead += BigInt(13000);
+                }
+                if (userOp.paymasterAndData !== '0x') {
+                    verificationGasLimitOverHead += BigInt(30000);
+                }
+                userOp.verificationGasLimit = '0x' + (BigInt(userOp.verificationGasLimit) + verificationGasLimitOverHead).toString(16);
+            }
+        } else {
+            // preVerificationGas overhead
+            {
+                // preVerificationGas overhead 1024 = 64bytes more ( 1024/16 = 64 )
+                userOp.preVerificationGas = '0x' + (BigInt(userOp.preVerificationGas) + BigInt(1024)).toString(16);
+            }
+
+            // verificationGasLimit overhead
+            {
+                let verificationGasLimitOverHead = BigInt(21000);
+                if (userOp.initCode !== '0x') {
+                    verificationGasLimitOverHead += BigInt(13000);
+                }
+                if (userOp.paymasterAndData !== '0x') {
+                    verificationGasLimitOverHead += BigInt(30000);
+                }
+                userOp.verificationGasLimit = '0x' + (BigInt(userOp.verificationGasLimit) + verificationGasLimitOverHead).toString(16);
+            }
         }
 
-        // verificationGasLimit overhead
-        {
-            let verificationGasLimitOverHead = BigInt(21000);
-            if (userOp.initCode !== '0x') {
-                verificationGasLimitOverHead += BigInt(13000);
-            }
-            if (userOp.paymasterAndData !== '0x') {
-                verificationGasLimitOverHead += BigInt(30000);
-            }
-            userOp.verificationGasLimit = '0x' + (BigInt(userOp.verificationGasLimit) + verificationGasLimitOverHead).toString(16);
-        }
+
     }
 }