+++ packages/soulwallet-abi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @soulwallet/abi
 
+## 0.1.2
+
+### Patch Changes
+
+- passkey version merged into soulwallet@npmjs.com
+
 ## 0.1.1
 
 ### Patch Changes