+++ packages/soulwallet-sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @soulwallet/sdk
 
+## 0.1.4
+
+### Patch Changes
+
+- Added L1KeyStore.getKeyHash(string[]) function
+
+## 0.1.3
+
+### Patch Changes
+
+- Allow caller to customize nonce
+
+## 0.1.2
+
+### Patch Changes
+
+- passkey version merged into soulwallet@npmjs.com
+- Updated dependencies
+  - @soulwallet/result@0.1.2
+  - @soulwallet/abi@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes